# ~3x performance improvement by optimizing code hotspots wrt regexes